+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaLocalizacionAbiertaRepositoryCustomImpl.java
@@ -56,14 +56,13 @@ public class TareaLocalizacionAbiertaRepositoryCustomImpl implements TareaLocali
     }
 
     @Override
-    public void trasladar(@NotNull TareaDto tareaDto, @NotNull List<Long> idTipoImporteVenta) {
+    public void trasladar(@NotNull TareaDto tareaDto, @NotNull List<Integer> idTipoImporteVenta) {
         MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ID_TAREA, tareaDto.getId());
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_IDS_TIPOS_DATO, idTipoImporteVenta);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_NUEVO_ID_TIPO_DATO,
                 TipoDatoEnum.VENTA_ONLINE_TRASLADADA.getId());
-        parameters.addValue(SqlPrimaryConstants.SQL_PARAM_MULTIPLICADOR,
-                SqlPrimaryConstants.SQL_VALUE_MULTIPLICADOR_POSITIVO);
+        parameters.addValue(SqlPrimaryConstants.SQL_PARAM_MULTIPLICADOR, SqlPrimaryConstants.SQL_VALUE_MULTIPLICADOR_POSITIVO);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ABIERTO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_TRUE);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_CERRADO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_FALSE);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_NUEVO_ACTIVO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_TRUE);
@@ -71,14 +70,13 @@ public class TareaLocalizacionAbiertaRepositoryCustomImpl implements TareaLocali
     }
 
     @Override
-    public void compensar(@NotNull TareaDto tareaDto, @NotNull List<Long> idTipoImporteVenta) {
+    public void compensar(@NotNull TareaDto tareaDto, @NotNull List<Integer> idTipoImporteVenta) {
         MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ID_TAREA, tareaDto.getId());
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_IDS_TIPOS_DATO, idTipoImporteVenta);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_NUEVO_ID_TIPO_DATO,
                 TipoDatoEnum.VENTA_ONLINE_COMPENSADA.getId());
-        parameters.addValue(SqlPrimaryConstants.SQL_PARAM_MULTIPLICADOR,
-                SqlPrimaryConstants.SQL_VALUE_MULTIPLICADOR_NEGATIVO);
+        parameters.addValue(SqlPrimaryConstants.SQL_PARAM_MULTIPLICADOR, SqlPrimaryConstants.SQL_VALUE_MULTIPLICADOR_NEGATIVO);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_ABIERTO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_TRUE);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_CERRADO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_FALSE);
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_NUEVO_ACTIVO, SqlPrimaryConstants.SQL_VALUE_BOOLEAN_TRUE);
@@ -86,7 +84,7 @@ public class TareaLocalizacionAbiertaRepositoryCustomImpl implements TareaLocali
     }
 
     @Override
-    public void saveCerrado(@NotNull TareaDto tareaDto, TrabajoDto trabajoDto, @NotNull List<Long> idTipoImporteVenta) {
+    public void saveCerrado(@NotNull TareaDto tareaDto, TrabajoDto trabajoDto, @NotNull List<Integer> idTipoImporteVenta) {
         MapSqlParameterSource parameters = new MapSqlParameterSource();
         parameters.addValue(SqlPrimaryConstants.SQL_PARAM_FECHA_INICIO,
                 TimeUtils.toDate(trabajoDto.getFechaInicioPeriodo()));
