+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaPersonaHistoricoRepositoryCustom.java
@@ -18,7 +18,7 @@ public interface TareaPersonaHistoricoRepositoryCustom {
             @NotNull @Positive String idOrigen);
 
     List<IdPersonaHistoricoDto> findIdPersonaHistoricoDtoByIdTareaAndIdOrigenAndTipoDatoInAmbito(@NotNull @Positive Long idTarea,
-            @NotBlank String idOrigen, @NotNull List<Long> idsTipoDato);
+            @NotBlank String idOrigen, @NotNull List<Integer> idsTipoDato);
 
     List<IdPersonaHistoricoDto> findIdPersonaHistoricoDtoByIdTareaAndIdOrigenInAmbito(@NotNull @Positive Long idTarea,
             @NotBlank String idOrigen);
